xen/arm32: entry: Consolidate DEFINE_TRAP_ENTRY*() macros
The only difference between the two macros DEFINE_TRAP_ENTRY() and
DEFINE_TRAP_ENTRY_NOIRQ() is the list of interrupts to be unmasked.

While the macros are fairly small today, it will be necessary to add the
same code twice in follow-up patches.

To prevent too much duplication, a new assembly macros is introduced to
generate the body of a trap.

This is part of XSA-303.

Reported-by: Julien Grall <Julien.Grall@arm.com>
Signed-off-by: Julien Grall <julien.grall@arm.com>
Acked-by: Stefano Stabellini <sstabellini@kernel.org>